// packages/features/src/products/errors/enum.ts
export enum ProductErrorCode {
  PRODUCT_SERVICE_UNAVAILABLE = 'PRODUCT_SERVICE_UNAVAILABLE',
  PRODUCT_CREATE_ERROR = 'PRODUCT_CREATE_ERROR',
  PRODUCT_DATABASE_ERROR = 'PRODUCT_DATABASE_ERROR',
  PRODUCT_ALREADY_EXISTS = 'PRODUCT_ALREADY_EXISTS',
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
}

export enum ProductVariantErrorCode {
  PRODUCT_VARIANT_SERVICE_UNAVAILABLE = 'PRODUCT_VARIANT_SERVICE_UNAVAILABLE',
  PRODUCT_VARIANT_CREATE_ERROR = 'PRODUCT_VARIANT_CREATE_ERROR',
  PRODUCT_VARIANT_DATABASE_ERROR = 'PRODUCT_VARIANT_DATABASE_ERROR',
  PRODUCT_VARIANT_ALREADY_EXISTS = 'PRODUCT_VARIANT_ALREADY_EXISTS',
  PRODUCT_VARIANT_NOT_FOUND = 'PRODUCT_VARIANT_NOT_FOUND',
}

export enum ProductImageErrorCode {
  PRODUCT_IMAGE_SERVICE_UNAVAILABLE = 'PRODUCT_IMAGE_SERVICE_UNAVAILABLE',
  PRODUCT_IMAGE_CREATE_ERROR = 'PRODUCT_IMAGE_CREATE_ERROR',
  PRODUCT_IMAGE_DATABASE_ERROR = 'PRODUCT_IMAGE_DATABASE_ERROR',
  PRODUCT_IMAGE_ALREADY_EXISTS = 'PRODUCT_IMAGE_ALREADY_EXISTS',
  PRODUCT_IMAGE_NOT_FOUND = 'PRODUCT_IMAGE_NOT_FOUND',
}

export enum ProductTagErrorCode {
  PRODUCT_TAG_SERVICE_UNAVAILABLE = 'PRODUCT_TAG_SERVICE_UNAVAILABLE',
  PRODUCT_TAG_CREATE_ERROR = 'PRODUCT_TAG_CREATE_ERROR',
  PRODUCT_TAG_DATABASE_ERROR = 'PRODUCT_TAG_DATABASE_ERROR',
  PRODUCT_TAG_ALREADY_EXISTS = 'PRODUCT_TAG_ALREADY_EXISTS',
  PRODUCT_TAG_NOT_FOUND = 'PRODUCT_TAG_NOT_FOUND',
}